refacotor serializers to use module and create specialized serializers for role and employee serializers
