+++ backend/api/urls.py
@@ -6,6 +6,6 @@ urlpatterns = [
     path("auth/signup/", views.signup),
     path("roles/", views.role_list, name='role_list'),
     path('roles/<int:role_id>', views.role_detail),
-    path('employees/', views.employee_list),
-    path('employees/<int:employee_id>', views.employee_detail)
+    # path('employees/', views.employee_list),
+    # path('employees/<int:employee_id>', views.employee_detail)
 ]
