+++ backend/api/views.py
@@ -1,14 +1,12 @@
-from rest_framework.decorators import api_view, authentication_classes, permission_classes
-from rest_framework.authentication import SessionAuthentication, TokenAuthentication
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.authtoken.models import Token
 
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.models import User
-from .models import Role, Employee
-from .serializers import UserSerializer, RoleSerializer, EmployeeSerializer
+from .models import Role
+from .serializers import UserSerializer, ReadRoleSerializer, ReadRoleWithEmployeeSerializer, WriteRoleSerializer
 
 
 @api_view(['POST'])
@@ -53,11 +51,11 @@ def role_list(request):
     """
     if request.method == 'GET':
         roles = Role.objects.all()
-        serializer = RoleSerializer(roles, many=True)
+        serializer = ReadRoleSerializer(roles, many=True)
         return Response(serializer.data)
 
     elif request.method == 'POST':
-        serializer = RoleSerializer(data=request.data)
+        serializer = WriteRoleSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -76,11 +74,11 @@ def role_detail(request, role_id):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
-        serializer = RoleSerializer(role)
+        serializer = ReadRoleWithEmployeeSerializer(role)
         return Response(serializer.data)
 
     elif request.method == 'PUT':
-        serializer = RoleSerializer(role, data=request.data)
+        serializer = WriteRoleSerializer(role, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -92,42 +90,42 @@ def role_detail(request, role_id):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-@api_view(['GET', 'POST'])
-def employee_list(request):
-    """
-    List all employees or create a new employee.
-    """
-    if request.method == 'GET':
-        employees = Employee.objects.all()
-        serializer = EmployeeSerializer(employees, many=True)
-        return Response(serializer.data)
+# @api_view(['GET', 'POST'])
+# def employee_list(request):
+#     """
+#     List all employees or create a new employee.
+#     """
+#     if request.method == 'GET':
+#         employees = Employee.objects.all()
+#         serializer = EmployeeSerializer(employees, many=True)
+#         return Response(serializer.data)
 
-    elif request.method == 'POST':
-        employee_serializer = EmployeeSerializer(data=request.data)
-        if employee_serializer.is_valid():
-            employee = employee_serializer.save()
+#     elif request.method == 'POST':
+#         employee_serializer = EmployeeSerializer(data=request.data)
+#         if employee_serializer.is_valid():
+#             employee = employee_serializer.save()
 
-            if 'roles' in request.data:
-                employee_roles = []
-                for employee_role in request.data['roles']:
-                    role = get_object_or_404(Role, role=employee_role['role'])
-                    employee_roles.append(role)
-                employee.roles.set(employee_roles)
-            return Response(employee_serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            return Response(employee_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+#             if 'roles' in request.data:
+#                 employee_roles = []
+#                 for employee_role in request.data['roles']:
+#                     role = get_object_or_404(Role, role=employee_role['role'])
+#                     employee_roles.append(role)
+#                 employee.roles.set(employee_roles)
+#             return Response(employee_serializer.data, status=status.HTTP_201_CREATED)
+#         else:
+#             return Response(employee_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET', 'PUT', 'DELETE'])
-def employee_detail(request, employee_id):
-    """
-    Retrieve, update, or delete employee
-    """
-    try:
-        employee = Employee.objects.get(id=employee_id)
-    except Employee.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+# @api_view(['GET', 'PUT', 'DELETE'])
+# def employee_detail(request, employee_id):
+#     """
+#     Retrieve, update, or delete employee
+#     """
+#     try:
+#         employee = Employee.objects.get(id=employee_id)
+#     except Employee.DoesNotExist:
+#         return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == 'GET':
-        serializer = EmployeeSerializer(employee)
-        return Response(serializer.data)
+#     if request.method == 'GET':
+#         serializer = EmployeeSerializer(employee)
+#         return Response(serializer.data)
